git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@5323 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/web/bean/spaces/DeleteSpaceDialog.java
@@ -86,7 +86,12 @@ public class DeleteSpaceDialog extends BaseDialogBean
          
          if (DELETE_ALL.equals(this.deleteMode))
          {
-            this.nodeService.deleteNode(node.getNodeRef());
+            NodeRef nodeRef = node.getNodeRef();
+            if (this.nodeService.exists(nodeRef))
+            {
+               // The node still exists
+               this.nodeService.deleteNode(node.getNodeRef());
+            }
          }
          else
          {
